pyproject.toml
[tool.poetry]
name = "dynamic-auto-scaling"
version = "0.1.0"
description = ""
authors = ["fruit-bird <github.ru3w9@slmail.me>"]
readme = "README.md"
packages = [{ include = "dynamic_auto_scaling", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
pydantic = "^2.7.0"
psutil = "^5.9.8"
asyncio = "^3.4.3"
adtk = "^0.6.2"
asyncpg = "^0.29.0"
requests = "^2.31.0"
pandas = "^2.2.2"
ipykernel = "^6.29.4"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
seaborn = "^0.13.2"
matplotlib = "^3.8.4"
streamlit = "^1.33.0"
python-dotenv = "^1.0.1"
altair = "^5.3.0"
mpld3 = "^0.5.10"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"

[tool.poetry.plugins.dotenv]
location = ".env"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
